Add title scene and fix car issues
