--- FroggerStopper/Assets/Scripts/StartButtonController.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class StartButtonController : MonoBehaviour
-{
-    private CarMovements[] allCarScripts;
-
-    void Start()
-    {
-        allCarScripts = FindObjectsOfType<CarMovements>();
-    }
-
-    public void StartButtonClicked()
-    {
-        foreach (CarMovements car in allCarScripts)
-        {
-            car.TriggerCarStart();
-        }
-    }
-}
