+++ FroggerStopper/ProjectSettings/TagManager.asset
@@ -11,7 +11,7 @@ TagManager:
   - Default
   - TransparentFX
   - Ignore Raycast
-  - 
+  - CarsLayer
   - Water
   - UI
   -
